完善了大部分的功能，，在6.24下午，实现了改卷的功能

+++ src/main/java/com/example/demo/entity/TopicSet.java
@@ -6,6 +6,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -26,11 +27,11 @@ public class TopicSet {
 
     private int tstatus;
 
-    private float taverage;
+    private BigDecimal taverage;
 
-    private float texcellentrate;
+    private BigDecimal texcellentrate;
 
-    private float tpassingrate;
+    private BigDecimal tpassingrate;
 
     private int tifcansee;
